% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{omp_set_num_threads}
\alias{omp_set_num_threads}
\alias{omp_get_num_threads}
\title{Set or get the number of OpenMP threads}
\usage{
omp_set_num_threads(num)

omp_get_num_threads()
}
\arguments{
\item{num}{The desired number of threads}
}
\value{
In the case of \code{omp_get_num_threads}, the current number of threads.
}
\description{
Set or get the number of threads used by OpenMP for parallel regions.
}
\details{
These functions mostly exist because of the CRAN check, which only allows 1 - 2 threads.
As there is a per-thread memory requirement for \code{RUVIII_C} and \code{RUVIII_C_Varying}, 
these functions can also be useful for controlling memory usage. However in that case, setting
environment variable OMP_NUM_THREADS also works.
}
