% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptideData}
\alias{peptideData}
\title{Peptide data from the cross-lab study of Collins et al (2017)}
\format{A data.frame containing peptide-level data, with peptides that were not detected coded as NA. All columns are peptides, except for \code{filename}, \code{site}, \code{mixture}, \code{grouping} and \code{Date}.}
\usage{
data(peptideData)
}
\description{
A dataset containing the peptide-level data from the cross-lab study of Collins et al (2017), originally published in Nature Communications. The data consists of a data.frame named \code{peptideData}, and a character vector sisPeptides, listing the peptides that should be analysed.
}
\keyword{datasets}
