% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVIII_C.R
\name{RUVIII_C}
\alias{RUVIII_C}
\title{RUV-III applied to complete rows}
\usage{
RUVIII_C(k, ruvInputData, M, toCorrect, filename, controls,
  withW = FALSE, batchSize = 1000)
}
\arguments{
\item{k}{The number of factors of unwanted variation to remove}

\item{ruvInputData}{The input data matrix. Must be a matrix, not a data.frame.}

\item{M}{The design matrix containing information about technical replicates}

\item{toCorrect}{The variables to correct using RUV-III-C}

\item{filename}{The intermediate file in which to save the results}

\item{controls}{The names of the control variables which are known to be constant across the observations.}

\item{withW}{Should we keep the estimate of the unwanted variation factors W, for each variable which is corrected?}

\item{batchSize}{How often should we write to the intermediate file? batchSize = 1000 implies that results are written to file every 1000 variables.}
}
\value{
If withW = FALSE, returns a matrix. If withW = TRUE, returns a list with entries named \code{newY} and \code{W}.
}
\description{
Apply RUV-III in the presence of missing values
}
\details{
This function applies RUV-III to the given dataset. However, it applies the RUV-III correction separately to every variable. If variable X is being corrected, we take the rows of the data matrix for which X is non-missing. RUV-III is then applied, and the corrected values of X is retained. The corrected values of all other variables are discarded. Note that when we subset the data matrix, other columns besides X will still have missing values. These values are replaced with zero in order to apply RUV-III.
}
\examples{
data(crossLab)
#Design matrix containing information about which runs are technical replicates of each other. 
#In this case, random pairings of mass-spec runs analysing the same sample, at different sites.
M <- model.matrix(~ grouping, data = peptideData)
#Get out the list of peptides, both HEK (control) and peptides of interest.
peptides <- setdiff(colnames(peptideData), c("filename", "site", "mixture", "Date", "grouping"))
#Reduce the data matrix to only the peptide data
onlyPeptideData <- data.matrix(peptideData[, peptides])
#All the human peptides are potential controls. That is, everything that's not an SIS peptides.
potentialControls <- setdiff(peptides, sisPeptides)
#But we want to use controls that are always found
potentialControlsAlwaysFound <- names(which(apply(onlyPeptideData[, potentialControls], 2, function(x) sum(is.na(x))) == 0))
#Because there are so many potential controls here, we only use 500. 
actualControls <- head(potentialControlsAlwaysFound, 500)
#Actually run correction
\dontrun{results <- RUVIII_C(k = 11, ruvInputData = onlyPeptideData, M = M, toCorrect = c(sisPeptides, actualControls), controls = actualControls, filename = "results.RData")}
}
