% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sisPeptides}
\alias{sisPeptides}
\title{List of stable isotope-labelled standard (SIS) peptides from the cross-lab study of Collins et al (2017)}
\format{
An object of class \code{character} of length 29.
}
\usage{
data(crossLab)
}
\description{
List of stable isotope-labelled standard (SIS) peptides from the cross-lab study of Collins et al (2017)
}
\keyword{datasets}
