% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVIII_C_Varying.R
\name{RUVIII_C_Varying}
\alias{RUVIII_C_Varying}
\title{RUV-III-C, varying controls}
\usage{
RUVIII_C_Varying(k, ruvInputData, M, toCorrect, filename,
  potentialControls, withW = FALSE, batchSize = 1000)
}
\arguments{
\item{k}{The number of factors of unwanted variation to remove}

\item{ruvInputData}{The input data matrix. Must be a matrix, not a data.frame. It should contain missing (NA) values, rather than zeros.}

\item{M}{The design matrix containing information about technical replicates. It should not contain an intercept term!}

\item{toCorrect}{The names of the variables to correct using RUV-III-C}

\item{filename}{The intermediate file in which to save the results.}

\item{potentialControls}{The names of the control variables which are known to be constant across the observations}

\item{withW}{Should we keep the estimate of the unwanted variation factors W, for each variable which is corrected?}

\item{batchSize}{How often should we write to the intermediate file? The default of 1000 implies that results are written to file every 1000 variables.}
}
\value{
If withW = FALSE, returns a matrix. If withW = TRUE, returns a list with entries named \code{newY} and \code{W}.
}
\description{
Apply RUV-III-C, a variation of RUV-III that only uses non-missing values
}
\details{
See the documentation of \link{RUVIII_C} for more information about the RUV-III-C method. This function is identical, except in this case the set of negative control variables actually used varies depending on the target variable to be normalized. Instead of putting in a list of negative control variables, the user specifies a list of potential negatve control variables.

When normalizing variable X, the algorithm begins by selecting the rows of the data matrix for which X is non-missing. Out of the potential negative control peptides, it selects those that are always non-missing across the selected subset. The standard version of RUV-III is then applied, similar to \link{RUVIII_C}.
}
